requirements.txt
python-slugify==3.0.2
requests==2.22.0
jsonschema==3.0.1
